+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Bug Fixes
 - [cronos#109](https://github.com/crypto-org-chain/cronos/issues/109) ibc transfer timeout too short
 
+### Features
+
+- [cronos#110](https://github.com/crypto-org-chain/cronos/pull/110) embed swagger doc ui
+
 *September 22, 2021*
 ## v0.5.4