+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [cronos#109](https://github.com/crypto-org-chain/cronos/issues/109) ibc transfer timeout too short
 - [tharsis#590](https://github.com/tharsis/ethermint/pull/590) fix export contract state in genesis and reimport
 - [cronos#123](https://github.com/crypto-org-chain/cronos/issues/123) fix ibc refund logic
+- [tharsis#617](https://github.com/tharsis/ethermint/pull/617) iterator on deeply nested cache contexts is extremely slow
+- [tharsis#615](https://github.com/tharsis/ethermint/pull/615) tx log attribtue value not parsable by some client
 
 ### Features